+++ tests/testthat/test-correct_colinvar.R
@@ -108,13 +108,65 @@ test_that("correct_colinvar PCA with projections", {
   terra::writeRaster(somevar, file.path(dir_sc[1], "somevar.tif"), overwrite=TRUE)
   terra::writeRaster(somevar, file.path(dir_sc[2], "somevar.tif"), overwrite=TRUE)
 
-  # Perform pearson collinearity control
+  # Perform PCA collinearity control
   var <-
     correct_colinvar(env_layer = somevar, method = "pca", proj = dirname(dir_sc[1]))
 
   expect_true(is.character(var$proj))
   unlink(dirname(dir_sc[1]), recursive = TRUE)
   unlink(gsub("projections", "Projection_PCA", dirname(dir_sc[1])), recursive = TRUE)
+})
+
+
+test_that("correct_colinvar PCA with different projection area", {
+  require(terra)
+  require(dplyr)
+  somevar <-
+    system.file("external/somevar.tif", package = "flexsdm")
+  somevar <- terra::rast(somevar)
+
+  # set seed
+  abies2 <- abies %>%
+    dplyr::select(x, y, pr_ab) %>%
+    dplyr::filter(pr_ab==1)
+
+  ca <- calib_area(abies2, x = "x", y = "y", method = c("mcp"), crs=crs(somevar))
+
+  # Perform PCA only with cell delimited by polygon used in restric_to_region
+  var <- correct_colinvar(env_layer = somevar ,
+                             method = c("pca"),
+                             maxcell = NULL,
+                             restric_to_region = ca,
+                             restric_pca_proj = FALSE)
+  expect_equal(length(var), 3)
+  expect_equal(class(var$env_layer)[1], "SpatRaster")
+  expect_equal(nrow(var$coefficients), 4)
+  expect_equal(nrow(var$cumulative_variance), 4)
+  expect_true(all(names(var) %in% c("env_layer", "coefficients", "cumulative_variance")))
+
+  # Perform and predicted PCA only with cell delimited by polygon used in restric_to_region
+  var <- correct_colinvar(env_layer = somevar ,
+                              method = c("pca"),
+                              maxcell = NULL,
+                              restric_to_region = ca,
+                              restric_pca_proj = TRUE)
+  expect_equal(length(var), 3)
+  expect_equal(class(var$env_layer)[1], "SpatRaster")
+  expect_true(ext(var$env_layer)[1]> (-310000))
+  expect_equal(nrow(var$coefficients), 4)
+  expect_equal(nrow(var$cumulative_variance), 4)
+  expect_true(all(names(var) %in% c("env_layer", "coefficients", "cumulative_variance")))
+
+
+  # Perform pearson collinearity control
+  var <-
+    correct_colinvar(env_layer = somevar, method = "pca")
+
+  expect_equal(length(var), 3)
+  expect_equal(class(var$env_layer)[1], "SpatRaster")
+  expect_equal(nrow(var$coefficients), 4)
+  expect_equal(nrow(var$cumulative_variance), 4)
+  expect_true(all(names(var) %in% c("env_layer", "coefficients", "cumulative_variance")))
 })
 
 test_that("correct_colinvar FA", {